Deployment/requirements.txt
streamlit>=1.1.0
pandas>=1.3.4,!=1.5.0,!=2.0.0
numpy>=1.21.0
Orange3==3.30.2
xgboost>=1.5.0
plotly>=5.18.0